fix: isolate token list test dependencies
