+++ packages/web/src/repositories/token/token-repository-impl.ts
@@ -12,9 +12,9 @@ import { NetworkClient } from "@common/clients/network-client";
 import { StorageClient } from "@common/clients/storage-client";
 import { CommonError } from "@common/errors";
 import { StorageKeyType } from "@common/values";
-import { customSort } from "@containers/select-token-container/SelectTokenContainer";
 import { TokenPriceModel } from "@models/token/token-price-model";
 import { TokenSearchLogModel } from "@models/token/token-search-log-model";
+import { customSort } from "@utils/token-sort";
 import { formatDisplayTokenSymbol } from "@utils/token-utils";
 import mockedExchangeRateGraph from "./mock/token-exchange-rate-graph.json";
 import { IBalancesByAddressResponse, IGrc20TransferHistoryResponse } from "./response/balance-by-address-response";
